refactor(interfaces): update type imports and introduce tokens module
- Changed type imports from 'chess-types' to 'tokens' for consistency across interfaces.
- Created a new 'tokens.ts' file to define low-level chess domain literals, improving type organization and clarity.
- Streamlined the index file by consolidating type exports, enhancing maintainability.
